+++ js/sending.js
@@ -1,0 +1,2 @@
+export {};
+
